Edit : More API edits and Tests
The new model base from Ollama is connected correctly to this system. The new bug now is the agent. which wouldn't work like the old agent. Guessing the output of the tools is too long and the model can't think of a way to parse it and give out error. Thus  trying to create a new agent or prompt to run it.