Incomplete & Feat : Formating and Dupe Mitre In Type Agent
Now TypeAgent will not run model when the log already have the model Mitre attack input. The formatting down stream are now align with the pydantic model input of the ContextAgent. But there is still some small issue in the Json string file. Will be corrected in the next commit