Incomplete : The Same bug persisted still
The bug where context agent called the LLM more than the time than the message send by NATS. right now suspected the problem of multiple agent running simultaneously checked and try to debug but unsure. NATS problem are mostly clear out but the configuration of the server might cause some problem but we can assume crossing out NATS as a source of bug.